Helpful python script to send test sequence to a real 2336 terminal
Term_2336.py offers routines for reading/writing to a serial-port
connected 2336.  It is hardwired to talk on COM13 (works for me) but
that is easily changed.  termtest_2336.py is a bunch of ad-hoc tests
that I've been using to verify terminal behavior so I can model it in
the emulator.