Cursor is 2 scanlines high, not 1
Unfortunately, a few unrelated things.
- The cursor is on the bottom two scanlines of the glyph
- The underline is on the bottom scanline of the glyph
- I whitened the green color to better match a real 2336 terminal
- I futzed around for hours trying to figure out why the rawbmp stuff
doesn't seem to work under windows.  Not sure if it was always like that
or something that happened since the last release.  It does work if I
create bitmaps with a depth of 24 (not 32, not -1), and use a
wxNativePixelData type.  For now I've just disabled it, and I'll debug
it later.
- a bunch of foo++ were turned into ++foo.