Checkpointing progress; MVP runs
The previous attempt at modeling the MXD terminal mux functionally has been abandoned.  Instead, I modified an intel 8080 emulator, by Alexander Demin, and emulate the MXD at an instruction level.

I added my own i8080 disassembler to help debug bringup of the MXD.

That required that I generalize the main emulator execution loop to do fine-grained switching between invoking the 2200 CPU and any peripheral 8080 CPUs.  This has resulted in a performance hit, some of which I'll try to claw back later.

The emulator event scheduler now uses 64b time math, which simplifies the bookkeeping and allows 1ns resolution, which is compatible with the time resolution used by the CPU clock ratio logic.

I've also generalized the "getIB5()" function to be just "getIB()" because the $GIO instruction actually requires it.  However, so far the behavior is unchanged, but at least the plumbing is in place to do it right.

A bug in the 2236 escape sequence processing was fixed, and now compressed runs of characters actually works.

The keyboard handling now changes key mappings for MXD mode so special function keys, and edit work.  Keyword mode is removed for MXD and is replaced by caps lock, just like a real 2236 terminal.