Smarter hard/warm reset handling
To trigger reset from the keyboard, the SHIFT key must be held down, to prevent accidental resets.

Rather than the CrtFrame directly calling the system reset, it now routes it through the keyboard handler. This is required for MXD operation, since resets are allowed only from terminal #1, and the rest get filtered out.