tweak: menu shortcut corrections
A previous check in changed reset to be Alt-Shift-R instead of just Alt-R.  That accidentally changed Alt-G (screen grab) and Alt-K (keyword mode) to also require Shift.  This fix reverts that.

Also, in 2236 terminal mode, there is no keyword mode, just caps lock.  The text on the menu shortcut was changed from "Keyword mode" to "Kaps lock" so the Alt-K was preserved.