small interface tweak
When setting up a new card in a slot, the address field should default to blank and force the user to pick an address.  However, because the base address of one of the choices for the 2236MXD card is 0x00, it matched intended "empty" value of 0 and was thus shown to be selected.

Now the not-yet-selected value is -1, so even MXD cards must have their addess set manually.