Adding cppcheck and fixing warnings
First sweep of the trivial warning fixes.

There are a number of suggestions to use std::fill and such which I will consider later.  There was on serious error found: the cpu2200vp ucode store was initialized (0 fillled) to the size of RAM instead of the size of the ucode store.