smart terminals get working save/restore
It hasn't mattered yet, but eventually the MXD terminal mux cards will support up to four terminals.  As a result, the ini file group string of "CRT-nn" is not sufficient.

That format is still used for dumb terminals, but serial terminals use the group string of MXD-nn-t, where nn is the MXD base address and t is the terminal number (1 to 4).