better terminal HEX(07) beep handling
If a bunch of HEX(07) are emitted in a row, the old way of telling windows to stop playing the beep in progress and then telling it to play it again was very choppy.

The new approach is to build a seamless wrap buffer of the desired tone, tell windows to loop it continuously, then use a wxTimer to indicate when to stop it.  If another HEX(07) arrives, it just resets the timer and doesn't interfere with the sound in progress.

Finally, my 2336 is one frequency but the dumb terminal controller schematics indicate there a 555 timer generates a lower beep.  The program now models the difference.