Fix a long-standing bug in card configuration editing
This bug exists in Wangemu 2.1.  When changing the system configuration, say slot #1 was the keyboard controller.  Say it is changed to be a disk controller.  Then pressing the card Config button to set disk controller options would result in no config dialog.

The old logic was not only wrong, but way overly complicated.