disable MXD at any addr other than 0x00
Running with more than one MXD results in a hang with the MVP OS, and I haven't debugged why that happens.  It is unweildy to run in that configuration anyway.

Also, back out the assertion that the shared cpu and sheduler pointers must be unique when reconfiguring the emulator because even though we've done a Destroy() on the old windows, wx doesn't actually destroy them until its event loop is empty.  Thus, when we build up the new windows, the old ones might linger, a bit, and the assumption that the refcount is 1 is false.  Thus, keep using shared_ptr.