clang-tidy cleanups, and more
I erred by not doing this in smaller bites, so if something got broken, it will be harder to figure out where.

Lots of changes suggested by clang-tidy, plus another round through cppcheck and MSVC++ lint. The most common change was the not use ints in a bool context and vice versa.

Made use of default member initialization in the class definition instead of in the constructor list.  I'm not sure why clang-tidy suggested it in only a few cases.

I also improved a corner case bug, that of doing script input in a 2200T when using an MXD for I/O.  A bit hacky but it will work in the 99% most common case.