First cut at the OSX build (xcode-free)
Adding a makefile for the osx build.  It is rough, but good enough for now.

I had to change a bit of C code to make osx clang happy. (the stuff in IoCard.{h,cpp})

The in-memory wxSound beep API is not supported on osx, so I created to .wav files for that platform.

The build.txt instructions have been updated with what I've done on Mojave to get things built, this time without xcode.

The program runs and is usable, but there are a number of regressions (some noted in todo.txt).