Fix very old bug in the population of the Disk menu
Until now, the disk menu assumed there were two drives per disk controller.  As a result, if there was only 1 it was possible to perform an action on the non-existent "R" drive, leading to a crash.  It was also not possible to use the menu to insert/remove disks to the secondary drives via the Disk menu.