fix osx hang after disk "inspect"
The routine in CrtFrame calls freezeCpu(true) before and freezeCpu(false) surrouind the call to the DiskFactory() dialog.  On OSX after the dialog was dismissed, the CPU remained frozen.

That was because modal dialogs are supposed to be ended by EndModal(), not Close().  Apparently by calling Close(), the thread of control returned to the main event loop without resuming the code in CrtFrame, and thus freezeCpu(false) was never called.