minor UI tweak
Previously, if the system configuration had an empty slot changed from empty to something then back to empty, the Revert button was still active.  Now Revert is disabled if there is nothing to revert in this case.

It happened because when it checked if old_cfg != new_cfg, it was still testing the card addr even if the card type was none.