bump up the limit on the number of scheduler events
One game taxed the (soft) limit on the number of pending timer events.  Every time the microcode retriggers the 27ms timer one-shot, it pushes an event in the scheduler queue.  It can do this multiple times in one 27ms window, resulting in multiple events; all of them but the last are zombies which just get removed when they retire, but in the mean time it increase the queue size.