convert a couple raw pointers to std::shared_ptr
Make the holder of the PrinterFrame and CrtFrame object use a smart pointer.

A std::unique_ptr would be even ideal, but that fails to compile because the full type of the object must be visible where it is used, and I don't want to expose the core emulator to all that GUI plumbing.

std::shared_ptr allows an abstract class to be used, so that is what I have used.