Restructure key event handling
I wanted to use Cmd-<blah> for handling keys that control the emulator (eg, menu shortcuts), and Ctrl-<blah> for mapping emulated keys (eg, Ctrl-P is PRINT).  But that was problematic with the old OnKeyDown routine.

I'm not sure why it was split between OnKeyDown and OnChar event handlers anyway. I assume that when it was first written it was necessary.  But using only OnChar works for MSW and OSX now.