Add information about OSX accented key menu feature
On OSX, there is a feature which causes a popup menu of accented character options when a key is held down long enough.  This isn't useful to the emulator, and in fact interferes with the expected behavior of long keypresses causing autorepeat.

Update the release notes to describe what is going on and how to work around it.

Update the application bundle with the proper "domain" tag to follow Apple's convention of listing the domain elements in reverse order (org.wang2200.www).