Switch over to templated help file generation
Originally I used TemplateToolkit ot generate the pc and osx versions of the help files. When I switched to using github and set up a new repository, I copied the html build products but not the generation logic.

I'm not deleting the build products and checking in the generation files.