Fix floppy drive mouse event handling on OSX
This has been an aggravating bug, but this change works around it. For some reason, wxStaticBitmap, which is used to represent the floppy disk icons on the statusbar, do not generate a mouse event when clicked on.

The workaround is to attach the handler on the statusbar and then manually check if the event position matches any of the static bitmaps; if there is a match, the event ID is converted to that of the floppy disk icon.