add CBS() handling to the disk controller
Previously, the disk controller ignored CBS strobes. Now the lsb is used to reset the disk controller state, which is accurate for phoenix-type drives, but not the floppy contollers. On the other hand, Wang BASIC doesn't issue CBS to the disk controller, just MVP OSs.

The disk controller now also tracks ABS-based and CBS-based disk hogging. Although the hog state is tracked, it is unused because there are no other devices competing for the disk drive.