Fix a bug in 2336 terminal fifo handling
Previously, the size of the queue of pending bytes was read outside the loop and that counter was decremented each time we consumed a byte in the loop. The error was that a call inside the loop could also consume bytes, making that loop counter out of sync.

Now the queue depth is checked each time through the loop.